requirements.txt
openai
opencc-python-reimplemented
pyannote.audio
torchvision
pydub